extracted logic from model to a trait
