+++ packages/panels/src/Auth/MultiFactor/App/Concerns/InteractsWithAppAuthentication.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Filament\Auth\MultiFactor\App\Concerns;
+
+trait InteractsWithAppAuthentication
+{
+    public function getAppAuthenticationSecret(): ?string
+    {
+        return $this->app_authentication_secret;
+    }
+
+    public function saveAppAuthenticationSecret(?string $secret): void
+    {
+        $this->app_authentication_secret = $secret;
+        $this->save();
+    }
+
+    public function getAppAuthenticationRecoveryCodes(): ?array
+    {
+        return $this->app_authentication_recovery_codes;
+    }
+}
