+++ packages/panels/src/Auth/MultiFactor/App/Concerns/InteractsWithAppAuthenticationRecovery.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Filament\Auth\MultiFactor\App\Concerns;
+
+trait InteractsWithAppAuthenticationRecovery
+{
+    public function saveAppAuthenticationRecoveryCodes(?array $codes): void
+    {
+        $this->app_authentication_recovery_codes = $codes;
+        $this->save();
+    }
+}
